updated ETH functionality with sending transaction
