--- Multy/Additional Libraries/Core Libs/Core/bip44_coin_types.h
@@ -1,184 +1,0 @@
-/* Copyright 2018 by Multy.io
- * Licensed under Multy.io license.
- *
- * See LICENSE for details
- */
-
-#ifndef BIP44_COIN_TYPES_H
-#define BIP44_COIN_TYPES_H
-
-#include <stddef.h>
-
-struct CoinType
-{
-    const size_t type;
-    const char* description;
-};
-
-// https://github.com/satoshilabs/slips/blob/master/slip-0044.md
-
-struct CoinType CointTypes[] = {
-        {0x80000000, "Bitcoin"},
-        {0x80000001, "Testnet (all coins)"},
-        {0x80000002, "Litecoin"},
-        {0x80000003, "Dogecoin"},
-        {0x80000004, "Reddcoin"},
-        {0x80000005, "Dash (ex Darkcoin)"},
-        {0x80000006, "Peercoin"},
-        {0x80000007, "Namecoin"},
-        {0x80000008, "Feathercoin"},
-        {0x80000009, "Counterparty"},
-        {0x8000000a, "Blackcoin"},
-        {0x8000000b, "NuShares"},
-        {0x8000000c, "NuBits"},
-        {0x8000000d, "Mazacoin"},
-        {0x8000000e, "Viacoin"},
-        {0x8000000f, "ClearingHouse"},
-        {0x80000010, "Rubycoin"},
-        {0x80000011, "Groestlcoin"},
-        {0x80000012, "Digitalcoin"},
-        {0x80000013, "Cannacoin"},
-        {0x80000014, "DigiByte"},
-        {0x80000015, "Open Assets"},
-        {0x80000016, "Monacoin"},
-        {0x80000017, "Clams"},
-        {0x80000018, "Primecoin"},
-        {0x80000019, "Neoscoin"},
-        {0x8000001a, "Jumbucks"},
-        {0x8000001b, "ziftrCOIN"},
-        {0x8000001c, "Vertcoin"},
-        {0x8000001d, "NXT"},
-        {0x8000001e, "Burst"},
-        {0x8000001f, "MonetaryUnit"},
-        {0x80000020, "Zoom"},
-        {0x80000021, "Vpncoin"},
-        {0x80000022, "Canada eCoin"},
-        {0x80000023, "ShadowCash"},
-        {0x80000024, "ParkByte"},
-        {0x80000025, "Pandacoin"},
-        {0x80000026, "StartCOIN"},
-        {0x80000027, "MOIN"},
-        {0x80000028, "Expanse"},
-        {0x80000029, "Einsteinium"},
-        {0x8000002a, "Decred"},
-        {0x8000002b, "NEM"},
-        {0x8000002c, "Particl"},
-        {0x8000002d, "Argentum"},
-        {0x8000002e, "Libertas"},
-        {0x8000002f, "Posw coin"},
-        {0x80000030, "Shreeji"},
-        {0x80000031, "Global Currency Reserve (GCRcoin)"},
-        {0x80000032, "Novacoin"},
-        {0x80000033, "Asiacoin"},
-        {0x80000034, "Bitcoindark"},
-        {0x80000035, "Dopecoin"},
-        {0x80000036, "Templecoin"},
-        {0x80000037, "AIB"},
-        {0x80000038, "EDRCoin"},
-        {0x80000039, "Syscoin"},
-        {0x8000003a, "Solarcoin"},
-        {0x8000003b, "Smileycoin"},
-        {0x8000003c, "Ether"},
-        {0x8000003d, "Ether Classic"},
-        {0x8000003e, "Pesobit"},
-        {0x8000003f, "Landcoin"},
-        {0x80000040, "Open Chain"},
-        {0x80000041, "Bitcoinplus"},
-        {0x80000042, "Internet of People"},
-        {0x80000043, "Nexus"},
-        {0x80000044, "InsaneCoin"},
-        {0x80000045, "OKCash"},
-        {0x80000046, "BritCoin"},
-        {0x80000047, "Compcoin"},
-        {0x80000048, "Crown"},
-        {0x80000049, "BelaCoin"},
-        {0x8000004a, "Compcoin"},
-        {0x8000004b, "FujiCoin"},
-        {0x8000004c, "LINK"},
-        {0x8000004d, "Verge"},
-        {0x8000004e, "Electronic Gulden"},
-        {0x8000004f, "ClubCoin"},
-        {0x80000050, "RichCoin"},
-        {0x80000051, "Potcoin"},
-        {0x80000052, "Quarkcoin"},
-        {0x80000053, "Terracoin"},
-        {0x80000054, "Gridcoin"},
-        {0x80000055, "Auroracoin"},
-        {0x80000056, "IXCoin"},
-        {0x80000057, "Gulden"},
-        {0x80000058, "BitBean"},
-        {0x80000059, "Bata"},
-        {0x8000005a, "Myriadcoin"},
-        {0x8000005b, "BitSend"},
-        {0x8000005c, "Unobtanium"},
-        {0x8000005d, "MasterTrader"},
-        {0x8000005e, "GoldBlocks"},
-        {0x8000005f, "Saham"},
-        {0x80000060, "Chronos"},
-        {0x80000061, "Ubiquoin"},
-        {0x80000062, "Evotion"},
-        {0x80000063, "SaveTheOcean"},
-        {0x80000064, "BigUp"},
-        {0x80000065, "GameCredits"},
-        {0x80000066, "Dollarcoins"},
-        {0x80000067, "Zayedcoin"},
-        {0x80000068, "Dubaicoin"},
-        {0x80000069, "Stratis"},
-        {0x8000006a, "Shilling"},
-        {0x8000006b, "MarsCoin"},
-        {0x8000006c, "Ubiq (UBQ)"},
-        {0x8000006d, "Pesetacoin"},
-        {0x8000006e, "Neurocoin"},
-        {0x8000006f, "ARK"},
-        {0x80000070, "UltimateSecureCashMain"},
-        {0x80000071, "Hempcoin"},
-        {0x80000072, "Linx"},
-        {0x80000073, "Ecoin"},
-        {0x80000074, "Denarius"},
-        {0x80000075, "Pinkcoin"},
-        {0x80000076, "PiggyCoin"},
-        {0x80000077, "Pivx (PIVX)"},
-        {0x80000078, "Flashcoin"},
-        {0x80000079, "Zencash"},
-        {0x8000007a, "Putincoin"},
-        {0x8000007b, "BitZeny"},
-        {0x8000007c, "Unify"},
-        {0x8000007d, "StealthCoin (XST)"},
-        {0x8000007e, "Breakout Coin (BRK)"},
-        {0x8000007f, "Vcash"},
-        {0x80000080, "Monero"},
-        {0x80000081, "Voxels"},
-        {0x80000082, "NavCoin"},
-        {0x80000083, "Factom Factoids"},
-        {0x80000084, "Factom Entry Credits"},
-        {0x80000085, "Zcash"},
-        {0x80000086, "Lisk"},
-        {0x80000087, "Steem"},
-        {0x80000088, "ZCoin"},
-        {0x80000089, "Rootstock"},
-        {0x8000008a, "Giftblock"},
-        {0x8000008b, "RealPointCoin"},
-        {0x8000008c, "LBRY Credits (LBC)"},
-        {0x8000008d, "Komodo (KMD)"},
-        {0x8000008e, "bisq Token (BSQ)"},
-        {0x8000008f, "Riecoin (RIC)"},
-        {0x80000090, "Ripple (XRP)"},
-        {0x80000091, "Bitcoin Cash"},
-        {0x80000092, "Neblio (NEBL)"},
-        {0x80000093, "ZClassic (ZCL)"},
-        {0x80000094, "Stellar Lumens"},
-        {0x80000095, "NoLimitCoin2 (NLC2)"},
-        //    {0x80000096, ""},
-        //    {0x80000097, ""},
-        //    {0x80000098, ""},
-        {0x80000099, "Bytom (BTM)"},
-        {0x8000009a, "Biocoin"},
-        {0x800000df, "AskCoin"},
-        {0x800000e0, "Smartcash"},
-        {0x8000014d, "MemCoin"},
-        {0x80000378, "NEO (NEO)"},
-        {0x80000539, "Defcoin"},
-        {0x800091be, "Rootstock Testnet"},
-};
-
-#endif // BIP44_COIN_TYPES_H
