+++ Multy/Extensions/BigInt+Extension.swift
@@ -4,12 +4,6 @@
 
 import Foundation
 
-extension _BigInt {
-    var btcString: String {
-        return self.toString().appendDelimeter(at: 8)
-    }
+struct BigIntSwift {
     
-    var ethString: String {
-        return self.toString().appendDelimeter(at: 18)
-    }
 }
