+++ Multy/Info.plist
@@ -32,9 +32,9 @@
 	<key>CFBundleVersion</key>
 	<string>34</string>
 	<key>GIT_BRANCH_NAME</key>
-	<string>MUL-1266_Slide_to_send</string>
+	<string>ETH_BETA_TEST</string>
 	<key>GIT_COMMIT_HASH</key>
-	<string>a27d2e9</string>
+	<string>5f8be96</string>
 	<key>GIT_TAG_NAME</key>
 	<string>Release-26-dirty</string>
 	<key>LSRequiresIPhoneOS</key>
