+++ Multy/Model/PONSO/EthereumGasInfo.swift
@@ -6,7 +6,7 @@ import UIKit
 
 class EthereumGasInfo: NSObject {
     
-    var gasPrice = 0
-    var gasLimit = 0
+    var gasPrice = 1000000000
+    var gasLimit = 21000
 
 }
